+++ R/enum.R
@@ -24,6 +24,7 @@ enumGetKey <- function(enum, value) {
 #' @export
 #'
 #' @examples
+#' library(ospsuite.utils)
 #' myEnum <- enum(c(a = "b"))
 #' myEnum <- enumPut("c", "d", myEnum)
 #' myEnum <- enumPut(c("c", "d", "g"), list(12, 2, "a"), myEnum, overwrite = TRUE)