+++ NEWS.md
@@ -1,6 +1,13 @@
-# esqlabsR 2.0.0
+# esqlabsR 3.0.0.9000
 
 BREAKING CHANGES
+
+- The package gains a new dependency: [`{ospsuite.utils}`](https://www.open-systems-pharmacology.org/OSPSuite.RUtils/).
+
+# esqlabsR 2.0.0
+
+- Maintenance and bug fixes.
+
 
 # esqlabsR 1.0.0